Multitenancy! :k8sordie:
This adds the ability to host multiple users and players in the same Kubernetes cluster, with the same oauther instance.

Players would ideally get access to create Jobs in their own namespace, and those pods run with a Service Account
that only has access to read secrets from their own namespace.

Save the token of each user in a new namespace, on the format "spotify-$USER_ID".

The refresh method scans all namespaces for oauth secrets.